removed duplication code into one common function
